Add user controller ready for other user resource endpoints and redirect /register to this
